consolidating changes _2023_11_04
